# perf(Symbol#inspect): Reuse onigmo regex for testing if symbol should be quoted