--- a/include/natalie/symbol_object.hpp
+++ b/include/natalie/symbol_object.hpp
@@ -66,6 +66,8 @@ public:
 
     const String &string() const { return m_name; }
 
+    bool should_be_quoted() const;
+
     virtual String dbg_inspect() const override;
 
     virtual void visit_children(Visitor &visitor) override {
@@ -85,6 +87,7 @@ public:
 
 private:
     inline static TM::Hashmap<const TM::String, SymbolObject *> s_symbols { TM::HashType::TMString, 1000 };
+    inline static regex_t *s_inspect_quote_regex { nullptr };
 
     SymbolObject(const String &name)
         : Object { Object::Type::Symbol, GlobalEnv::the()->Symbol() }