This speeds up `Symbol#inspect` by quite a lot by reusing the regex object.

For this silly benchmark:

```ruby
10_000.times do |i|
  i.to_s.to_sym.inspect
end
```

...it produces about a 15x speedup:

```
→ hyperfine ./before ./after
Benchmark 1: ./before
  Time (mean ± σ):     788.0 ms ±  17.1 ms    [User: 783.7 ms, System: 4.1 ms]
  Range (min … max):   757.5 ms … 815.0 ms    10 runs
 
Benchmark 2: ./after
  Time (mean ± σ):      47.6 ms ±   0.7 ms    [User: 45.0 ms, System: 2.6 ms]
  Range (min … max):    46.2 ms …  49.6 ms    61 runs
 
Summary
  './after' ran
   16.55 ± 0.44 times faster than './before'
```